requirements.txt
Django==1.9
django-bootstrap-pagination==1.6.1
django-sortable-column==0.2.2
psycopg2==2.6.1
reportlab==3.3.0
neo4j-driver==1.0.2
djangorestframework==3.4.6
requests==2.11.1
django-favicon==0.1.2
